+++ src/main/java/ninja/mbedded/ninjaterm/util/comPort/ComPortFactory.java
@@ -1,0 +1,12 @@
+package ninja.mbedded.ninjaterm.util.comPort;
+
+/**
+ * Created by gbmhu on 2016-10-28.
+ */
+public class ComPortFactory {
+
+    public ComPort create() {
+        return new ComPort();
+    }
+
+}