+++ changelog.md
@@ -1,8 +1,12 @@
-vX.X.X
+v0.8.6
 ------
 
 - Added script to install openjfx before main Linux installer runs, closes #186.
 - Added installation steps for all platforms, closes #184.
+- Fixed bug where 'Open' button on TX/RX tab is enabled when no COM ports can be found, closes #185.
+- Removed manual buffer size limitations on totalNewLineParserOutput, closes #178.
+- Fixed bug where WebView data pane labels where not being setup correctly on start-up, closes #187.
+- Build system changed from Maven to Gradle.
 
 v0.8.5
 ------